Refactor Foundation Models: split ViewModels, inject dynamic backend, remove Smart Planner, and modernize UI
