+++ firebaseai/FirebaseAIExample/Features/AppleFoundationModels/Screens/HybridAIScreen.swift
@@ -1,4 +1,4 @@
-// Copyright 2025 Google LLC
+// Copyright 2026 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,11 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Foundation
+import SwiftUI
 
-public enum BackendOption: String, CaseIterable, Identifiable {
-  case googleAI = "Google AI"
-  case vertexAI = "Firebase Vertex AI"
+@available(iOS 27.0, *)
+struct HybridAIScreen: View {
+  @StateObject private var viewModel: HybridAIViewModel
 
-  public var id: String { rawValue }
+  init(backendType: BackendOption) {
+    _viewModel = StateObject(wrappedValue: HybridAIViewModel(backendType: backendType))
+  }
+
+  var body: some View {
+    FoundationModelsContainer(viewModel: viewModel, title: "Hybrid AI") { vm in
+      HybridAIView(viewModel: vm)
+    }
+  }
 }
