+++ firebaseai/FirebaseAIExample/Features/AppleFoundationModels/ViewModels/FoundationModelsBaseViewModel.swift
@@ -1,0 +1,58 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import Foundation
+import SwiftUI
+import Combine
+import FoundationModels
+import FirebaseAILogic
+
+public enum ModelPreference: String, CaseIterable, Identifiable {
+  case auto = "Auto (Local First)"
+  case cloud = "Cloud Only"
+
+  public var id: String { rawValue }
+}
+
+@available(iOS 27.0, *)
+@MainActor
+public class FoundationModelsBaseViewModel: ObservableObject {
+  @Published public var inProgress = false
+  @Published public var error: Error?
+  @Published public var modelPreference: ModelPreference = .auto
+  @Published public var isUsingLocalModel: Bool = false
+
+  internal var activeTask: Task<Void, Never>?
+  public let backendType: BackendOption
+
+  public init(backendType: BackendOption) {
+    self.backendType = backendType
+  }
+
+  public func stopActiveTask() {
+    activeTask?.cancel()
+    activeTask = nil
+    inProgress = false
+  }
+
+  internal func getFirebaseAI() -> FirebaseAI {
+    switch backendType {
+    case .googleAI:
+      return FirebaseAI.firebaseAI(backend: .googleAI())
+    case .vertexAI:
+      // Using "global" as location for Foundation Models fallback compatibility
+      return FirebaseAI.firebaseAI(backend: .vertexAI(location: "global"))
+    }
+  }
+}
